Move batch_generator to utils
