+++ src/pkg/caendr/caendr/utils/data.py
@@ -111,3 +111,19 @@ def get_file_format(file_ext, valid_formats=None):
 
   # If none matched, return None
   return None
+
+
+def batch_generator(g, batch_size=1000000):
+  '''
+    Split a generator into a generator of generators, which produce the same sequence when taken together.
+    Useful for managing RAM when bulk inserting mappings into a table.
+  '''
+  def _inner(top):
+    yield top
+    for i, x in enumerate(g, start=1):
+      yield x
+      if i % (batch_size - 1) == 0:
+        return
+
+  for top in g:
+    yield _inner(top)
